% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{bootstrap_variable}
\alias{bootstrap_variable}
\title{Bootstrap Resampling with Variable Sample Sizes (R version)}
\usage{
bootstrap_variable(data, min_size, max_size, n_bootstrap)
}
\arguments{
\item{data}{Original data vector to bootstrap from}

\item{min_size}{Minimum bootstrap sample size}

\item{max_size}{Maximum bootstrap sample size}

\item{n_bootstrap}{Number of bootstrap samples to generate}
}
\value{
A list of bootstrap samples, each with a different random size
}
\description{
Performs bootstrap resampling where each bootstrap sample has a randomly
  determined size. This demonstrates another natural use case for dynamic vector growth.
}
