% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bootstrap_variable_cpp4r}
\alias{bootstrap_variable_cpp4r}
\title{Bootstrap Resampling with Variable Sample Sizes}
\usage{
bootstrap_variable_cpp4r(data, min_size, max_size, n_bootstrap)
}
\arguments{
\item{data}{Original data vector}

\item{min_size}{Minimum bootstrap sample size}

\item{max_size}{Maximum bootstrap sample size}

\item{n_bootstrap}{Number of bootstrap samples}
}
\description{
Another natural use case for push_back - bootstrap sampling where
  sample sizes can vary based on some condition
}
