% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4rsampling-package.R
\docType{package}
\name{cpp4rsampling-package}
\alias{cpp4rsampling}
\alias{cpp4rsampling-package}
\title{cpp4rsampling: Bivariate Gibbs Sampler functions for benchmarks}
\description{
ADD DESCRIPTION. TWO OR MORE LINES
}
\seealso{
Useful links:
\itemize{
  \item \url{https://WEBSITE.COM}
  \item Report bugs at \url{https://github.com/USERNAME/PKGNAME/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

}
\keyword{internal}
