% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{rejection_sampling}
\alias{rejection_sampling}
\title{Rejection Sampling from Truncated Normal Distribution (R version)}
\usage{
rejection_sampling(n_samples, mu = 0, sigma = 1, lower = -2, upper = 2)
}
\arguments{
\item{n_samples}{Number of accepted samples desired}

\item{mu}{Mean of the normal distribution. Default is 0.}

\item{sigma}{Standard deviation of the normal distribution. Default is 1.}

\item{lower}{Lower truncation bound. Default is -2.}

\item{upper}{Upper truncation bound. Default is 2.}
}
\value{
A numeric vector of accepted samples from the truncated normal distribution
}
\description{
Generates samples from a truncated normal distribution using rejection sampling.
  This R implementation demonstrates the same algorithm as the cpp4r version but will be slower
  due to the overhead of growing vectors in R.
}
