% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{rejection_sampling_cpp4r}
\alias{rejection_sampling_cpp4r}
\title{Rejection Sampling from Truncated Normal Distribution}
\usage{
rejection_sampling_cpp4r(n_samples, mu, sigma, lower, upper)
}
\arguments{
\item{n_samples}{Number of accepted samples desired}

\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{lower}{Lower truncation bound}

\item{upper}{Upper truncation bound}
}
\description{
Demonstrates dynamic vector growth with rejection sampling - you don't know 
  how many samples you need to generate to get n_samples accepted values
}
